fix(llvm) : 把全局变量传到 llvmProgram 中
