+++ example.sy
@@ -1,0 +1,3 @@
+int main() {
+    return 0;
+}
