+++ utils/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod errors;
 pub use errors::*;
+pub mod init_value_item;
+pub use init_value_item::*;
 
 pub fn fatal_error(str: &str) {
 	eprintln!("{}: {}", console::style("fatal error").bold().red(), str);
